server/requirements.txt
pyjwt
fastapi
tinydb
python-multipart